requirements.txt
streamlit
textblob
pandas
altair
vaderSentiment
PyPDF2
python-docx
